Set `text-align: right` in rtl language
Some frameworks will set `text-align: left`, for example: `bootstrap`
in the test site. So we use `dir` attribute to mark the direction of
banner and preferences dialog. If `dir="rtl"`, we set
`text-align: right`.

The `dir` attribute will be set when the `ConsentControl` instance is
initialized. It will only affect the element itself and its children. In
our case, we set `dir` attribute in the banner and preferences dialog,
so the body (contents) of banner and preferences dialog will be
affected. And it will not influence other parts of the website.