Add functions to set themes
Some styles can be determined by other styles, if they are not passed.
For example, "cookieModal-background-color" is the style under
preferences dialog that creates smoking background to cover whole page.
It can be determined by "preferences-background-color", which is the
style for preferences dialog, and an additional parameter to indicate alpha channel. This is the second parts in `setCertainTheme`.